server.js/requirements.txt
Flask==2.3.2
Flask_Cors==4.0.0
langchain==0.1.4
langchain_core==0.1.17
langchain_google_genai==0.0.6
pandas==2.0.3
pymongo==4.6.1
python-dotenv==1.0.1
